requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
pydantic>=2
python-dotenv
Pillow>=10.0.0
playwright

